+++ src/front/js/index.js
@@ -2,11 +2,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-//include your index.scss file into the bundle
-import "../styles/index.css";
+// include your styles into the webpack bundle
+// import "../styles/styles.css";
 
 //import your own components
-import Layout from "./layout";
+import Layout from './layout.js';
 
 //render your react application
 ReactDOM.render(<Layout />, document.querySelector("#app"));
