+++ src/front/js/layout.js
@@ -4,12 +4,11 @@ import ScrollToTop from "./component/scrollToTop";
 import { BackendURL } from "./component/backendURL";
 
 import { Home } from "./pages/home";
-import { Demo } from "./pages/demo";
-import { Single } from "./pages/single";
 import injectContext from "./store/appContext";
 
 import { Navbar } from "./component/navbar";
 import { Footer } from "./component/footer";
+import { LoginSignUp } from "./LoginSignUp";
 
 //create your first component
 const Layout = () => {
@@ -26,8 +25,7 @@ const Layout = () => {
                     <Navbar />
                     <Routes>
                         <Route element={<Home />} path="/" />
-                        <Route element={<Demo />} path="/demo" />
-                        <Route element={<Single />} path="/single/:theid" />
+                        <Route element={<LoginSignUp />} path="/" />
                         <Route element={<h1>Not found!</h1>} />
                     </Routes>
                     <Footer />
