--- src/front/jsx/app.jsx
@@ -1,14 +1,0 @@
-import React, {useState} from "react";
-import {xxxxx} from "./xxxxxx";
-
-
-
-
-//create your first component
-const app = () => {
-
-return(
-		<LoginSignUp/>
-	)
-}
-export default app;
